requirements.txt
llvmlite==0.43.0
numba==0.60.0
numpy==2.0.2
pillow==10.4.0
